backend/requirements.txt
celery==4.0.0
Django==1.10.2
django-celery-results==1.0.0
django-cors-headers==1.2.2
django-oauth-toolkit==0.10.0
djangorestframework==3.5.1
mysqlclient==1.3.9
python-social-auth==0.2.21
Unipath==1.1
